% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxau.R
\docType{data}
\name{maxau}
\alias{maxau}
\title{Annual suspended sediment concentration and flow data, River Rhine}
\format{
a time series object of class "mts"

\itemize{
\item s. first column, suspended sediment concentration
\item Q. second column, average discharge
}
}
\source{
Bundesanstalt für Gewässerkunde, Koblenz, Deutschland
(Federal Institute of Hydrology, Koblenz, Germany)
}
\usage{
data(maxau)
}
\description{
Annual time series of average suspended sediment
 concentration (s) in mg/l and average discharge (Q) in m^3 / s at
 the River Rhine, 1965.1-2009.1
}
\examples{
data(maxau)
plot(maxau)
}
\keyword{datasets}
