% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cptest.R
\name{plot.cptest}
\alias{plot.cptest}
\title{Plotting cptest-objects}
\usage{
\method{plot}{cptest}(x, ...)
}
\arguments{
\item{x}{an object of class "cptest"}

\item{\ldots}{further arguments, currently ignored}
}
\description{
Plotting method for objects inheriting from class "cptest"
}
\examples{
data(Nile)
(out <- br.test(Nile))
par(mfrow=c(2,1))
plot(Nile) ; plot(out)

}
