% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.smk.test.R
\name{summary.smktest}
\alias{summary.smktest}
\title{Object summaries}
\usage{
\method{summary}{smktest}(object, ...)
}
\arguments{
\item{object}{an object of class \code{smktest}}

\item{\ldots}{further arguments, currently ignored}
}
\description{
Generic function "summary" for objects of class \code{smktest}.
}
